Fix paths for tests, remove component tests, rename files, clean up code
