+++ src/model/schedule.js
